Move wasm.minimal.annotation to wasm.annotation.
